% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bcb.R
\name{download_bcb}
\alias{download_bcb}
\title{Download data from Banco Central do Brasil (BCB) API}
\usage{
download_bcb(codigo, data_inicial, data_final)
}
\arguments{
\item{codigo}{A character string representing the series code for the desired data. The series
code can be found on the BCB website (https://www.bcb.gov.br) or in the BCB API
documentation (https://www3.bcb.gov.br/sgspub/localizarseries/localizarSeries.do?method=prepararTelaLocalizarSeries).}

\item{data_inicial}{A character string representing the start date of the desired data in the format
"dd-mm-yyyy".}

\item{data_final}{A character string representing the end date of the desired data in the format "dd-mm-yyyy".}
}
\value{
A tibble with two columns: date and valor.
}
\description{
This function downloads time-series data from the Banco Central do Brasil (BCB) API
for a specified series code, start date, and end date. The function returns a tibble
with two columns: date and value. Optionally, the function can also write the data to an
Excel file.
}
\examples{
download_bcb("4389", "01-01-2020", "31-12-2020")
}
