% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_api_error.R
\name{handle_api_error}
\alias{handle_api_error}
\title{Lidar com erros da API}
\usage{
handle_api_error(res)
}
\arguments{
\item{res}{Resposta da API}
}
\value{
Retorna um erro com uma mensagem explicativa
}
\description{
Função para lidar com possíveis erros na resposta da API do BCB.
}
