% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformar-data.R
\name{transformar_data}
\alias{transformar_data}
\title{Transform a date to BR format}
\usage{
transformar_data(data)
}
\arguments{
\item{data}{date to transform}
}
\value{
a date in a character form
}
\description{
Transform a date to BR format
}
\examples{
transformar_data("24-05-2023")
}
